整合处理报错

+++ carcode/system/carmotor.h
@@ -15,7 +15,7 @@
 void Motor_Init(void);
 void PWM_Init(void);
 void Set_Pwm(int motor_left,int motor_right);   //设置pwm
-int PWM_Limit(int IN,int max,int min);   //设置pwm最大最小值
+float PWM_Limit(float IN,float max,float min);   //设置pwm最大最小值
 int myabs(int a);    //绝对值函数
